misc progress

+++ kernel/src/interrupts.c
@@ -1,9 +1,9 @@
 #include "types.h"
 
-void interrupt_handler(u32 raw_irq_number) {
+void interrupt_handler(usize raw_irq_number) {
     return;
 }
 
-void exception_handler(u32 raw_cause) {
+void exception_handler(usize raw_cause) {
     return;
 }
